Fix response header check
If a webserver or reverse proxy itself also sets the required headers,
the check will yield a false negative since the header will be set twice
(once from nextcloud, once from the webserver).

From
https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/getResponseHeader:
`If there are multiple response headers with the same name, then their
values are returned as a single concatenated string, where each value is
separated from the previous one by a pair of comma and space`.

So in case the header is set twice, `getResponseHeader` will return
something like `SAMEORIGIN, SAMEORIGIN` and the equals check will fail
and therefore result in a false negative. Using String.indexOf() to
search for a substring will prevent this.