Show correct status in update notification
Sometimes it could be the case that after a channel change the last updated
timestamp is reset, but the check for updates is not yet completed. In this
case the update notification showed "You are up to date" and the date was
"1970-01-01 00:00:00". With this fix the state is properly shown.

Signed-off-by: Morris Jobke <hey@morrisjobke.de>