Fix loading icon inside upload button in Theming app
Pull request #5969 fixed the vertical alignment of loading icon and
status message in Theming app, but in doing so it broke the loading icon
inside the upload button.

When an image is being uploaded the upload icon is replaced by a loading
icon, so the loading icon and the upload icon have to share their CSS
rules, but only in that specific case.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>