Ignore clicks on login button icon so they are received by the button
Visually, the icon positioned on the login button is part of the login
button, so clicking on the icon should be the same as clicking on the
button. As the icon is a different element this commit sets the icon to
ignore mouse events so they are received and handled by the login button
instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>