Do not iterate over the files
As "singleFileUpload" is used the "add" callback (which in turn calls
"addFileToUpload") will always be called with a single file. Therefore
there is no need to iterate over the files (and it is not done in the
other callbacks either, so this aligns the code with the rest of the
callbacks).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>