Fix incorrect quota in upload to shared folder by PROPFIND
This fixes issue #5437. It replaces reading the folder's free space from
the #free_space HTML element by a WebDAV PROPFIND request. This is
relevant if uploading by dragging onto a shared folder's icon in the web
interface, as #free_space reports the user's quota whereas PROPFIND
returns the quota of the share's owner in that case.