Unify appearance of avatars for undefined and unknown users
When calling the jQuery avatar plugin with a user that did not exist
(that is, users for which "/avatar/{user}/{size}" return a JSON response
with an empty "displayname" value) "?" on a grey background was shown.
However, if the jQuery avatar plugin was called with an undefined
JavaScript value then "?" was shown on a bluish background. This commit
unifies both cases to use the grey background.

The unit tests were also modified to ensure that the grey background is
used in both cases.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>