Remove "cursor: pointer" from quota cell
The quota is set through a dropdown, and this dropdown is only shown
when clicking on the select element itself. Therefore, the cell should
not show a pointer when hovering on it but outside the select element.

Note that although the removed "&.quota" selector also set a "width"
declaration that declaration was later overriden by
"#userlist td.quota", so it was not needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>