Fix resending mail notifications when searching by e-mail address
Besides the user name and the full name, a sharee of type
"SHARE_TYPE_USER" can also be found by looking for an e-mail address.
Ironically, in this case the "hasEmailAddress" property was not set on
the returned data, so the frontend did not offer the option to send
again the mail notification.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>