Fix contacts menu for IE11
IE11 triggers an 'input' event whenever an input is focussed
or loses focus. Thus this causes an endless loading loop as soon
as the view is re-rendered. To prevent this, this remembers the
previous search term and ignores events where the term has not
changed.

Fixes https://github.com/nextcloud/server/issues/5281

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>