Updated OpenGraph title to display folder name
When sharing a folder via Facebook, WhatsApp & Co (any platform that makes use of OpenGraph) the display title was "Nextcloud - a safe home for all your data" and the description was the actual name of the folder. This should not be the case, so I changed the display title to display the folder name and the description now displays the form OpenGraph title.