Fix input field not disabled while a comment is being sent
Since the change of the text area to a content editable div the input
field was no longer disabled while a new comment was being sent. It was
caused by still trying to disable the div using the "disabled" property,
which works only on real input fields; when using a content editable div
the way to disable it is by setting "contenteditable" to "false".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>