Disable elements while a comment is being deleted
When a comment is being deleted the "disabled" class is added to the
comment div, which causes it to look disabled. However, the input
elements and the content editable div were not truly disabled, and thus
it was still possible to interact with them. This commit ensures that
they are properly disabled while the comment is being deleted.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>