Disable contacts menu for mentions to current user in comment messages
The contacts menu is not shown for avatars and user names in the author
row if they represent the current user. For consistency, and because the
contacts menu provides no value when shown for the current user, this
commit also disables the contacts menu for mentions to the current user.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>