Only in case of $currentAccess the array uses the id as index
Otherwise its a normal string[] with the user ids, in that
case the array_merge did it's job just fine, apart from it
not being deduplicated.
The array+array is only needed when the user id is the key,
so integer only user ids are kept as they are instead of being
reindexed.

Regression from 3820d6883dffcaa49deb1054ae0f32ab3d3e39b1

Signed-off-by: Joas Schilling <coding@schilljs.com>