Allows adding of hex color to the theme-color
Fixes #7158. Adds a # on the color if missing.

Increased maxlength, added hash:true for jscolor, and adding a # if not present on the change event.

Since the input element now allows a hex code, changed values to hexcode.

In addition, added a function to get RGB array from hex or rgb values. Calling it in both methods and using it to perform comparison.

Also changed the way we were determining whether the jscolor component had loaded.

Changed the control to use data-jscolor rather than defining opts in the class.

Signed-off-by: Abijeet <abijeetpatro@gmail.com>