Limit the draggable area of files from the Files app
When a file from the Files app was dragged there were no limits to where
it can be dragged. Due to this it can disappear behind the header and it
can be dragged infinitely to the right of the list of files (a scroll
bar appears allowing the user to keep dragging the file to the right).

As far as I know, currently files from the Files app can be dropped only
on the file list or the breadcrumbs, so it is not necessary to be able
to drag them outside that area. Therefore, the dragged file is now
forced to stay inside the "#app-content-files" element, which ensures
that it will not disappear behind the header and prevents it from being
infinitely dragged to the right.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>