Do no run SyncJob in cron
Fixes #7184

The SyncJob can be very resource intensive. Since it requests all users
on the system to create the system addressbook. In order to do this it
creates a vcard for every user and updates the addressbook.

There is no need for this job since the proper signals are emitted and
handled in the carddav backend to update the addressbook live.

Worst comes to worst there is always the occ command to bring the
address book in sync again.

Signed-off-by: Roeland Jago Douma <roeland@famdouma.nl>