Fix file list visible behind the multiselect header
Some parts of the file list contents (file name and actions) had a
higher z-index than the file list multiselect header. That header is
fixed in place, so when the file list contents were scrolled and those
parts with a higher z-index overlapped the multiselect header they were
fully visible. Now the z-index for the multiselect header has a higher
value (the same used in the controls bar) to ensure that the contents
are shown behind the header.

Fixes #7540

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>