Fix popover menu not closing in user settings
A popover menu is displayed when its element has the "open" CSS class.
That class is added when clicking on the menu toggle, but it was removed
only when clicking again on the toggle, so the popover menu in user
settings could be closed only by clicking again on the menu toggle. Now
the "open" CSS class is removed too when clicking on the body, either
directly or through event propagation.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>